refactor: LottoGameMachine에 대한 추상화 해제 및 klint 에러 해결
