--- src/main/kotlin/lotto/domain/shop/machine/LottoGameMachine.kt
@@ -1,8 +1,0 @@
-package lotto.domain.shop.machine
-
-import lotto.domain.shop.LottoGame
-
-interface LottoGameMachine {
-
-    fun create(): LottoGame
-}
