+++ src/main/kotlin/main.kt
@@ -1,6 +1,6 @@
 import lotto.domain.analysis.RealLottoResultAnalyst
 import lotto.domain.shop.RealLottoShop
-import lotto.domain.shop.machine.RealLottoGameMachine
+import lotto.domain.shop.LottoGameMachine
 import lotto.view.LottoController
 import lotto.view.RealLottoInputView
 import lotto.view.RealLottoResultView
@@ -11,7 +11,7 @@ fun main() {
         lottoInputView = RealLottoInputView(),
         lottoResultView = RealLottoResultView(),
         lottoShop = RealLottoShop(
-            lottoGameMachine = RealLottoGameMachine(
+            lottoGameMachine = LottoGameMachine(
                 shuffler = RandomShuffler(),
             ),
         ),
