+++ src/test/kotlin/lotto/domain/shop/RealLottoGameMachineTest.kt
@@ -1,4 +1,4 @@
-package lotto.domain.shop.machine
+package lotto.domain.shop
 
 import shffule.MockLottoNumberDescendingSortShuffler
 import io.kotest.core.spec.style.StringSpec
@@ -7,7 +7,7 @@ import io.kotest.matchers.shouldBe
 class RealLottoGameMachineTest : StringSpec({
 
     "로또 게임에 담긴 로또 번호 목록은 오름차순 정렬이 되어있다" {
-        val lottoGame = RealLottoGameMachine(MockLottoNumberDescendingSortShuffler()).create()
+        val lottoGame = LottoGameMachine(MockLottoNumberDescendingSortShuffler()).create()
         lottoGame.lottoNumbers.value shouldBe lottoGame.lottoNumbers.value.sorted()
     }
 })
