--- src/test/kotlin/lotto/domain/shop/machine/MockLottoGameMachine.kt
@@ -1,12 +1,0 @@
-package lotto.domain.shop.machine
-
-import lotto.domain.shop.LottoGame
-
-class MockLottoGameMachine(
-    private val lottoGameCreateDelegator: () -> LottoGame
-) : LottoGameMachine {
-
-    override fun create(): LottoGame {
-        return lottoGameCreateDelegator()
-    }
-}
